requirements.txt
# CathayBot 依赖

# NoneBot2 核心
nonebot2>=2.3.0
nonebot-adapter-onebot>=2.4.0

# 驱动器
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# 配置
pyyaml>=6.0
pydantic>=2.0.0

# 数据库
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0  # PostgreSQL 支持

# WebUI (可选)
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]>=1.7.4  # 密码哈希
python-multipart>=0.0.9  # 表单解析

# 工具
httpx>=0.27.0
aiofiles>=23.0.0

# Redis 缓存
redis>=5.0.0

# 图片渲染 (统计/状态图片输出)
nonebot-plugin-htmlrender>=0.3.0
jinja2>=3.1.0

# 开发依赖 (可选)
# black>=24.0.0
# isort>=5.13.0
# mypy>=1.8.0
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
